requirements.txt
# Core packages
fastapi==0.73.0
pydantic==1.9.0
uvicorn==0.17.0
spacy==3.2.1
medspacy==0.2.0.0
dkpro-cassis==0.7.0
negspacy==1.0.2
textdescriptives==1.0.6

#Manually added dependencies to troubleshoot installation issues
attrs==21.4.0
deprecation==2.1.0
sortedcontainers==2.4.0
lxml==4.7.1
more-itertools==8.12.0
toposort==1.7
pyphen==0.12.0
pandas==1.4.1
pytz==2021.3
python-dateutil==2.8.2

click==8.0.3
asgiref==3.5.0
h11==0.13.0
starlette==0.17.1
typing-extensions==4.0.1
anyio==3.5.0
sniffio==1.2.0
thinc==8.0.13
numpy==1.22.2
wasabi==0.9.0
srsly==2.4.2
catalogue==2.0.6
cymem==2.0.6
preshed==3.0.6
murmurhash==1.0.6
blis==0.7.5
langcodes==3.3.0
packaging==21.3
pyparsing==3.0.7
tqdm==4.62.3
typer==0.4.0
requests==2.27.1
urllib3==1.26.8
charset-normalizer==2.0.11
certifi==2021.10.8
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
Cython==0.29.27
pysbd==0.3.4
